Allow queuing deletes

+++ src/Searchable.php
@@ -80,7 +80,13 @@ trait Searchable
             return;
         }
 
-        return $models->first()->searchableUsing()->delete($models);
+        if (! config('scout.queue')) {
+            return $models->first()->searchableUsing()->delete($models);
+        }
+
+        dispatch((new RemoveFromSearch($models))
+                ->onQueue($models->first()->syncWithSearchUsingQueue())
+                ->onConnection($models->first()->syncWithSearchUsing()));
     }
 
     /**
